Added all code for fire + made fire prefab and added some to scene + changed all damage and hp vars to floats

+++ Rapid Fire Interactive's FPS Mid-Term/Assets/Scripts/gameManager.cs
@@ -146,7 +146,7 @@ public class gameManager : MonoBehaviour {
         }
     }
 
-    public void updateBossBar(Image _bossHealthBar, float bossHP, int _health)
+    public void updateBossBar(Image _bossHealthBar, float bossHP, float _health)
     {
         // Check if the player has entered the boss room
         if (bossRoom.instance.getBossFightState())
