+++ Rapid Fire Interactive's FPS Mid-Term/Assets/Scripts/upgradeMenu.cs
@@ -28,7 +28,7 @@ public class upgradeMenu : MonoBehaviour {
     int damageRank = 0;
     int speedRank = 0;
     int staminaRank = 0;
-    int HP;
+    float HP;
     float damageMod = 1;
     float speed;
     int stamina;
